setup middleware logger
